Refactor Babamul API routes and update enrichment schemas
- Move top-level auth routes to babamul/mod.rs
- Updated routes for consistency
- Removed schemars, added`utoipa::ToSchema` in various structs for LSST and ZTF alert properties.
- Updated test cases to reflect changes in API routes and ensure proper functionality for alert processing and retrieval.
- Cleaned up test data management for alerts and objects in the database.
- rename multiple objects endpoint to alerts, since that is what they are retrieving (not object level data)
