+++ Cargo.toml
@@ -33,7 +33,6 @@ rand = "0.9.2"
 rdkafka = { version = "0.38.0", features = ["gssapi", "ssl", "tracing", "sasl"] }
 redis = { version = "0.32.7", features = ["tokio-comp"] }
 reqwest = { version = "0.12.24", features = ["json", "stream"] }
-schemars = "1.0.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.145"
 serde_with = "3.15.0"
