+++ src/api/docs.rs
@@ -89,12 +89,14 @@ pub struct ApiDoc;
         description = "The Public REST API for Babamul."
     ),
     paths(
-        routes::babamul::users::post_babamul_signup,
-        routes::babamul::users::post_babamul_activate,
-        routes::babamul::users::post_babamul_auth,
-        routes::babamul::users::get_babamul_profile,
-        routes::babamul::schemas::get_babamul_schema,
-        routes::babamul::surveys::get_object,
+        routes::babamul::post_babamul_signup,
+        routes::babamul::post_babamul_activate,
+        routes::babamul::post_babamul_auth,
+        routes::babamul::get_babamul_profile,
+        routes::babamul::surveys::schemas::get_babamul_schema,
+        routes::babamul::surveys::objects::get_object,
+        routes::babamul::surveys::alerts::get_alert_cutouts,
+        routes::babamul::surveys::alerts::get_alerts,
     ),
     security(
         ("babamul_jwt_token" = [])
