+++ src/utils/lightcurves.rs
@@ -1,7 +1,6 @@
 use apache_avro_derive::AvroSchema;
 use apache_avro_macros::serdavro;
 use mongodb::bson::doc;
-use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, skip_serializing_none};
 
@@ -31,7 +30,7 @@ pub fn diffmaglim2fluxerr(diffmaglim: f32, zp: f32) -> f32 {
 }
 
 #[apache_avro_macros::serdavro]
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize, Eq, Hash, schemars::JsonSchema)]
+#[derive(Debug, PartialEq, Clone, Deserialize, Serialize, Eq, Hash, utoipa::ToSchema)]
 pub enum Band {
     #[serde(rename = "g")]
     G,
@@ -47,7 +46,7 @@ pub enum Band {
     U,
 }
 
-#[derive(Debug, Clone, serde::Deserialize, serde::Serialize, AvroSchema, JsonSchema)]
+#[derive(Debug, Clone, serde::Deserialize, serde::Serialize, AvroSchema)]
 pub struct PhotometryMag {
     #[serde(alias = "jd")]
     pub time: f64,
@@ -60,7 +59,7 @@ pub struct PhotometryMag {
 
 #[serde_as]
 #[skip_serializing_none]
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize, AvroSchema, JsonSchema)]
+#[derive(Debug, PartialEq, Clone, Deserialize, Serialize, AvroSchema, utoipa::ToSchema)]
 pub struct BandRateProperties {
     pub rate: f32,
     pub r_squared: f32,
@@ -72,7 +71,7 @@ pub struct BandRateProperties {
 // (this needs to be fixed in the apache_avro-related crates)
 // #[serde_as]
 // #[skip_serializing_none]
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize, AvroSchema, JsonSchema)]
+#[derive(Debug, PartialEq, Clone, Deserialize, Serialize, AvroSchema, utoipa::ToSchema)]
 pub struct BandProperties {
     pub peak_jd: f64,
     pub peak_mag: f32,
@@ -87,7 +86,7 @@ pub struct BandProperties {
 // #[serde_as]
 // #[skip_serializing_none]
 #[serdavro]
-#[derive(Debug, PartialEq, Clone, Deserialize, Serialize, JsonSchema, Default)]
+#[derive(Debug, PartialEq, Clone, Deserialize, Serialize, Default, utoipa::ToSchema)]
 pub struct PerBandProperties {
     pub g: Option<BandProperties>,
     pub r: Option<BandProperties>,
