+++ tests/api/test_babamul.rs
@@ -3,6 +3,7 @@ mod tests {
     use actix_web::http::StatusCode;
     use actix_web::middleware::from_fn;
     use actix_web::{test, web, App};
+    use boom::alert::{AlertWorker, ProcessAlertStatus};
     use boom::api::auth::{babamul_auth_middleware, get_test_auth};
     use boom::api::db::get_test_db_api;
     use boom::api::email::EmailService;
@@ -10,12 +11,17 @@ mod tests {
     use boom::api::routes;
     use boom::api::test_utils::{read_json_response, read_str_response};
     use boom::conf::{load_dotenv, AppConfig};
+    use boom::enrichment::{EnrichmentWorker, LsstEnrichmentWorker};
+    use boom::utils::enums::Survey;
+    use boom::utils::testing::{
+        drop_alert_from_collections, lsst_alert_worker, AlertRandomizer, TEST_CONFIG_FILE,
+    };
     use mongodb::bson::doc;
     use mongodb::Database;
 
     /// Helper struct to manage test user lifecycle
     struct TestUser {
-        pub user: boom::api::routes::babamul::users::BabamulUser,
+        pub user: boom::api::routes::babamul::BabamulUser,
         pub token: String,
         database: Database,
     }
@@ -29,7 +35,7 @@ mod tests {
         ) -> Self {
             let test_email = format!("test+{}@babamul.example.com", email_suffix);
             let babamul_users_collection: mongodb::Collection<
-                boom::api::routes::babamul::users::BabamulUser,
+                boom::api::routes::babamul::BabamulUser,
             > = database.collection("babamul_users");
 
             // Clean up any existing user with this email
@@ -38,7 +44,7 @@ mod tests {
                 .await
                 .ok();
 
-            let test_user = boom::api::routes::babamul::users::BabamulUser {
+            let test_user = boom::api::routes::babamul::BabamulUser {
                 id: uuid::Uuid::new_v4().to_string(),
                 username: "testuser".to_string(),
                 email: test_email.clone(),
@@ -55,7 +61,7 @@ mod tests {
 
             // Create JWT token for test user
             let (token, _) =
-                boom::api::routes::babamul::users::create_babamul_jwt(auth_app_data, &test_user.id)
+                boom::api::routes::babamul::create_babamul_jwt(auth_app_data, &test_user.id)
                     .await
                     .expect("Failed to create JWT");
 
@@ -75,7 +81,7 @@ mod tests {
 
             tokio::spawn(async move {
                 let babamul_users_collection: mongodb::Collection<
-                    boom::api::routes::babamul::users::BabamulUser,
+                    boom::api::routes::babamul::BabamulUser,
                 > = database.collection("babamul_users");
                 babamul_users_collection
                     .delete_one(doc! { "_id": &user_id })
@@ -107,7 +113,7 @@ mod tests {
                 .app_data(web::Data::new(database.clone()))
                 .app_data(web::Data::new(auth_app_data.clone()))
                 .app_data(web::Data::new(EmailService::new()))
-                .service(routes::babamul::users::post_babamul_signup),
+                .service(routes::babamul::post_babamul_signup),
         )
         .await;
 
@@ -148,9 +154,8 @@ mod tests {
         );
 
         // Verify the user was created in the database
-        let babamul_users_collection: mongodb::Collection<
-            boom::api::routes::babamul::users::BabamulUser,
-        > = database.collection("babamul_users");
+        let babamul_users_collection: mongodb::Collection<boom::api::routes::babamul::BabamulUser> =
+            database.collection("babamul_users");
         let user = babamul_users_collection
             .find_one(doc! { "email": &test_email })
             .await
@@ -226,9 +231,9 @@ mod tests {
                 .app_data(web::Data::new(database.clone()))
                 .app_data(web::Data::new(auth_app_data.clone()))
                 .app_data(web::Data::new(EmailService::new()))
-                .service(routes::babamul::users::post_babamul_signup)
-                .service(routes::babamul::users::post_babamul_activate)
-                .service(routes::babamul::users::post_babamul_auth),
+                .service(routes::babamul::post_babamul_signup)
+                .service(routes::babamul::post_babamul_activate)
+                .service(routes::babamul::post_babamul_auth),
         )
         .await;
 
@@ -247,9 +252,8 @@ mod tests {
         assert_eq!(resp.status(), StatusCode::OK);
 
         // Get the activation code from the database
-        let babamul_users_collection: mongodb::Collection<
-            boom::api::routes::babamul::users::BabamulUser,
-        > = database.collection("babamul_users");
+        let babamul_users_collection: mongodb::Collection<boom::api::routes::babamul::BabamulUser> =
+            database.collection("babamul_users");
         let user = babamul_users_collection
             .find_one(doc! { "email": &test_email })
             .await
@@ -384,7 +388,7 @@ mod tests {
                 .app_data(web::Data::new(database.clone()))
                 .app_data(web::Data::new(auth_app_data.clone()))
                 .app_data(web::Data::new(EmailService::new()))
-                .service(routes::babamul::users::post_babamul_signup),
+                .service(routes::babamul::post_babamul_signup),
         )
         .await;
 
@@ -411,18 +415,18 @@ mod tests {
     #[actix_rt::test]
     async fn test_get_babamul_schema() {
         load_dotenv();
-        let babamul_schemas = boom::api::routes::babamul::schemas::BabamulAvroSchemas::new();
+        let babamul_schemas = boom::api::routes::babamul::surveys::BabamulAvroSchemas::new();
 
         let app = test::init_service(
             App::new()
                 .app_data(web::Data::new(babamul_schemas))
-                .service(routes::babamul::schemas::get_babamul_schema),
+                .service(routes::babamul::surveys::get_babamul_schema),
         )
         .await;
 
         // ZTF schema
         let req = test::TestRequest::get()
-            .uri("/babamul/schema/ztf")
+            .uri("/babamul/surveys/ztf/schemas")
             .to_request();
 
         let resp = test::call_service(&app, req).await;
@@ -441,7 +445,7 @@ mod tests {
 
         // LSST schema
         let req = test::TestRequest::get()
-            .uri("/babamul/schema/lsst")
+            .uri("/babamul/surveys/lsst/schemas")
             .to_request();
 
         let resp = test::call_service(&app, req).await;
@@ -460,7 +464,7 @@ mod tests {
 
         // Invalid survey
         let req = test::TestRequest::get()
-            .uri("/babamul/schema/invalid_survey")
+            .uri("/babamul/surveys/invalid_survey/schemas")
             .to_request();
 
         let resp = test::call_service(&app, req).await;
@@ -471,9 +475,9 @@ mod tests {
         );
     }
 
-    /// Test GET /babamul/surveys/{survey_name}/objects/cutouts/{candid} success case
+    /// Test GET /babamul/surveys/{survey_name}/objects/{candid}/cutouts success case
     #[actix_rt::test]
-    async fn test_get_object_cutouts() {
+    async fn test_get_alert_cutouts() {
         load_dotenv();
         let database: Database = get_test_db_api().await;
         let auth_app_data = get_test_auth(&database).await.unwrap();
@@ -508,13 +512,13 @@ mod tests {
                 .app_data(web::Data::new(database.clone()))
                 .app_data(web::Data::new(auth_app_data.clone()))
                 .wrap(from_fn(babamul_auth_middleware))
-                .service(routes::babamul::surveys::get_object_cutouts),
+                .service(routes::babamul::surveys::get_alert_cutouts),
         )
         .await;
 
         let req = test::TestRequest::get()
             .uri(&format!(
-                "/babamul/surveys/ztf/objects/cutouts/{}",
+                "/babamul/surveys/ztf/alerts/{}/cutouts",
                 test_candid
             ))
             .insert_header(("Authorization", format!("Bearer {}", test_user.token)))
@@ -546,7 +550,7 @@ mod tests {
 
         // Test retrieval of non-existent candid
         let req = test::TestRequest::get()
-            .uri("/babamul/surveys/ztf/objects/cutouts/8888888888")
+            .uri("/babamul/surveys/ztf/alerts/8888888888/cutouts")
             .insert_header(("Authorization", format!("Bearer {}", test_user.token)))
             .to_request();
 
@@ -558,9 +562,9 @@ mod tests {
         );
     }
 
-    /// Test GET /babamul/surveys/{survey_name}/objects
+    /// Test GET /babamul/surveys/{survey_name}/alerts
     #[actix_rt::test]
-    async fn test_get_objects() {
+    async fn test_get_lsst_alerts() {
         load_dotenv();
         let database: Database = get_test_db_api().await;
         let auth_app_data = get_test_auth(&database).await.unwrap();
@@ -578,41 +582,56 @@ mod tests {
                 .app_data(web::Data::new(database.clone()))
                 .app_data(web::Data::new(auth_app_data.clone()))
                 .wrap(from_fn(babamul_auth_middleware))
-                .service(routes::babamul::surveys::get_objects),
+                .service(routes::babamul::surveys::get_alerts),
         )
         .await;
 
+        let mut alert_worker = lsst_alert_worker().await;
+        let (candid, object_id, _, _, bytes_content) =
+            AlertRandomizer::new_randomized(Survey::Lsst)
+                .ra(180.0)
+                .dec(0.0)
+                .get()
+                .await;
+        let status = alert_worker.process_alert(&bytes_content).await.unwrap();
+        assert_eq!(status, ProcessAlertStatus::Added(candid));
+        let mut enrichment_worker = LsstEnrichmentWorker::new(TEST_CONFIG_FILE).await.unwrap();
+        let result = enrichment_worker.process_alerts(&[candid]).await;
+        assert!(result.is_ok(), "Enrichment failed: {:?}", result.err());
         // Query with cone search and magnitude filters
         let req = test::TestRequest::get()
-            .uri("/babamul/surveys/ZTF/objects?ra=180.0&dec=0.0&radius_arcsec=60&min_magpsf=15&max_magpsf=20")
-            .insert_header(("Authorization", format!("Bearer {}", test_user.token)))
-            .to_request();
-
-        let resp = test::call_service(&app, req).await;
-        // Should either succeed (200) or have a server error due to query structure
-        // The important thing is that it passes authentication
-        assert!(
-            resp.status() == StatusCode::OK || resp.status() == StatusCode::INTERNAL_SERVER_ERROR,
-            "Request should not be unauthorized"
-        );
-
-        // Query with non-ZTF survey should return error
-        let req = test::TestRequest::get()
-            .uri("/babamul/surveys/LSST/objects")
+            .uri("/babamul/surveys/lsst/alerts?ra=180.0&dec=0.0&radius_arcsec=60&min_magpsf=11&max_magpsf=26")
             .insert_header(("Authorization", format!("Bearer {}", test_user.token)))
             .to_request();
 
         let resp = test::call_service(&app, req).await;
         assert_eq!(
             resp.status(),
-            StatusCode::BAD_REQUEST,
-            "Only ZTF survey is supported for get_objects"
+            StatusCode::OK,
+            "Should successfully retrieve alerts (error: {})",
+            read_str_response(resp).await
         );
+        let body = read_json_response(resp).await;
+        let alerts = body["data"].as_array().unwrap();
+        assert!(
+            !alerts.is_empty(),
+            "Response should contain at least one alert"
+        );
+        assert!(
+            alerts
+                .iter()
+                .any(|alert| alert["objectId"].as_str().unwrap() == object_id
+                    && alert["candid"].as_i64().unwrap() == candid),
+            "Response should contain the inserted alert"
+        );
+
+        // Clean up
+        drop_alert_from_collections(candid, "LSST").await.unwrap();
     }
 
     /// Test GET /babamul/surveys/{survey_name}/objects/{object_id}
     #[actix_rt::test]
-    async fn test_get_object() {
+    async fn test_get_lsst_object() {
         load_dotenv();
         let database: Database = get_test_db_api().await;
         let auth_app_data = get_test_auth(&database).await.unwrap();
@@ -625,87 +644,14 @@ mod tests {
         )
         .await;
 
-        // Insert test data with unique IDs
-        let suffix = unique_suffix();
-        let object_id = format!("ZTF_test_obj_{}", suffix);
-        let candid: i64 = suffix.parse::<i64>().unwrap() + 77777;
-
-        // Insert alert with candidate and classifications
-        let alerts_collection = database.collection::<mongodb::bson::Document>("ZTF_alerts");
-
-        // Clean up any existing data
-        alerts_collection
-            .delete_one(doc! { "_id": candid })
-            .await
-            .ok();
-
-        let alert_doc = mongodb::bson::doc! {
-            "_id": candid,
-            "objectId": &object_id,
-            "candidate": {
-                "jd": 2460000.5,
-                "ra": 180.0,
-                "dec": 0.0,
-            },
-            "properties": {
-                "rock": false,
-            },
-            "classifications": {
-                "drb": 0.9,
-            }
-        };
-        alerts_collection
-            .insert_one(&alert_doc)
-            .await
-            .expect("Failed to insert alert");
-
-        // Insert cutouts
-        let cutouts_collection =
-            database.collection::<mongodb::bson::Document>("ZTF_alerts_cutouts");
-        cutouts_collection
-            .delete_one(doc! { "_id": candid })
-            .await
-            .ok();
-
-        let cutout_doc = mongodb::bson::doc! {
-            "_id": candid,
-            "cutoutScience": mongodb::bson::Binary {
-                subtype: mongodb::bson::spec::BinarySubtype::Generic,
-                bytes: vec![1, 2, 3, 4, 5],
-            },
-            "cutoutTemplate": mongodb::bson::Binary {
-                subtype: mongodb::bson::spec::BinarySubtype::Generic,
-                bytes: vec![6, 7, 8, 9, 10],
-            },
-            "cutoutDifference": mongodb::bson::Binary {
-                subtype: mongodb::bson::spec::BinarySubtype::Generic,
-                bytes: vec![11, 12, 13, 14, 15],
-            },
-        };
-        cutouts_collection
-            .insert_one(&cutout_doc)
-            .await
-            .expect("Failed to insert cutout");
-
-        // Insert aux data
-        let aux_collection = database.collection::<mongodb::bson::Document>("ZTF_alerts_aux");
-        aux_collection
-            .delete_one(doc! { "_id": &object_id })
-            .await
-            .ok();
-
-        let aux_doc = mongodb::bson::doc! {
-            "_id": &object_id,
-            "prv_candidates": [],
-            "prv_nondetections": [],
-            "fp_hists": [],
-            "cross_matches": {},
-            "aliases": {},
-        };
-        aux_collection
-            .insert_one(&aux_doc)
-            .await
-            .expect("Failed to insert aux data");
+        let mut alert_worker = lsst_alert_worker().await;
+        let (candid, object_id, _, _, bytes_content) =
+            AlertRandomizer::new_randomized(Survey::Lsst).get().await;
+        let status = alert_worker.process_alert(&bytes_content).await.unwrap();
+        assert_eq!(status, ProcessAlertStatus::Added(candid));
+        let mut enrichment_worker = LsstEnrichmentWorker::new(TEST_CONFIG_FILE).await.unwrap();
+        let result = enrichment_worker.process_alerts(&[candid]).await;
+        assert!(result.is_ok(), "Enrichment failed: {:?}", result.err());
 
         let app = test::init_service(
             App::new()
@@ -717,7 +663,7 @@ mod tests {
         .await;
 
         let req = test::TestRequest::get()
-            .uri(&format!("/babamul/surveys/ztf/objects/{}", &object_id))
+            .uri(&format!("/babamul/surveys/lsst/objects/{}", &object_id))
             .insert_header(("Authorization", format!("Bearer {}", test_user.token)))
             .to_request();
 
@@ -725,7 +671,8 @@ mod tests {
         assert_eq!(
             resp.status(),
             StatusCode::OK,
-            "Should successfully retrieve object"
+            "Should successfully retrieve object (error: {})",
+            read_str_response(resp).await
         );
 
         let body = read_json_response(resp).await;
@@ -744,22 +691,11 @@ mod tests {
         );
 
         // Clean up
-        alerts_collection
-            .delete_one(doc! { "_id": candid })
-            .await
-            .expect("Failed to delete alert");
-        cutouts_collection
-            .delete_one(doc! { "_id": candid })
-            .await
-            .expect("Failed to delete cutout");
-        aux_collection
-            .delete_one(doc! { "_id": &object_id })
-            .await
-            .expect("Failed to delete aux");
+        drop_alert_from_collections(candid, "LSST").await.unwrap();
 
         // Test retrieval of non-existent object
         let req = test::TestRequest::get()
-            .uri("/babamul/surveys/ztf/objects/nonexistent_object")
+            .uri("/babamul/surveys/lsst/objects/nonexistent_object")
             .insert_header(("Authorization", format!("Bearer {}", test_user.token)))
             .to_request();
